fix: delete pir_bfv hardware
