+++ pir_bfv/CMakeLists.txt
@@ -1,16 +1,14 @@
 set(USE_HARDWARE OFF)
-if(USE_HARDWARE)
-    add_library(pir_of_bfv pir.h pir.cpp pir_client_hardware.h pir_client_hardware.cpp pir_server_hardware.h
-            pir_server_hardware.cpp)
-    add_definitions(-DPIR_USE_HARDWARE)
-    #add_definitions(-DDEBUG)
+
+if(PIR_USE_HARDWARE)
     message(STATUS "pir hardware version")
+    add_definitions(-DPIR_USE_HARDWARE)
 else()
-    add_library(pir_of_bfv pir.h pir.cpp pir_client.h pir_client.cpp pir_server.h
-            pir_server.cpp)
     message(STATUS "pir software version")
 endif()
 
+add_library(pir_of_bfv pir.h pir.cpp pir_client.h pir_client.cpp pir_server.h
+        pir_server.cpp)
 target_link_libraries(pir_of_bfv PUBLIC poseidon_shared)
 
 add_executable(bfv_pir main.cpp)
