+++ pir_bfv/pir_client.cpp
@@ -11,9 +11,10 @@ namespace pir
 PIRClient::PIRClient(const ParametersLiteral &enc_params, const PirParams &pir_params)
     : enc_params_(enc_params), pir_params_(pir_params)
 {
-    PoseidonFactory::get_instance()->set_device_type(DEVICE_SOFTWARE);
-    context_ = std::make_shared<PoseidonContext>(
-        PoseidonFactory::get_instance()->create_poseidon_context(enc_params));
+    PoseidonFactory::get_instance()->set_device_type(DEVICE_HARDWARE);
+    context_ =
+        std::make_shared<PoseidonContext>(PoseidonFactory::get_instance()->create_poseidon_context(
+            enc_params));
 
     keygen_ = std::make_unique<KeyGenerator>(*context_);
 
@@ -32,8 +33,6 @@ PIRClient::PIRClient(const ParametersLiteral &enc_params, const PirParams &pir_p
 
     decryptor_ = std::make_unique<Decryptor>(*context_, secret_key);
     evaluator_ = PoseidonFactory::get_instance()->create_bfv_evaluator(*context_);
-    // evaluator_ = std::make_unique<EvaluatorBfvBase>(
-    //     PoseidonFactory::get_instance()->create_bfv_evaluator(*context_));
     encoder_ = std::make_unique<BatchEncoder>(*context_);
 }
 
@@ -237,18 +236,7 @@ Plaintext PIRClient::decode_reply(PirReply &reply)
         {
             Plaintext ptxt;
             decryptor_->decrypt(temp[j], ptxt);
-
-#ifdef DEBUG
-            std::cout << "Client: reply noise budget = "
-                      << decryptor_->invariant_noise_budget(temp[j]) << std::endl;
-#endif
-
             tempplain.push_back(ptxt);
-
-#ifdef DEBUG
-            std::cout << "recursion level : " << i << " noise budget :  ";
-            std::cout << decryptor_->invariant_noise_budget(temp[j]) << std::endl;
-#endif
 
             if ((j + 1) % (exp_ratio * ciphertext_size) == 0 && j > 0)
             {
@@ -283,19 +271,6 @@ Plaintext PIRClient::decode_reply(PirReply &reply)
 
 GaloisKeys PIRClient::generate_galois_keys()
 {
-    // Generate the Galois keys needed for coeff_select.
-    // vector<uint32_t> galois_elts;
-    // int N = enc_params_.degree();
-    // int logN = get_power_of_two(N);
-
-    // // std::cout << "printing galois elements...";
-    // for (int i = 0; i < logN; i++)
-    // {
-    //     galois_elts.push_back((N + exponentiate_uint(2, i)) / exponentiate_uint(2, i));
-    //     //#ifdef DEBUG
-    //     // std::cout << galois_elts.back() << ", ";
-    //     //#endif
-    // }
     GaloisKeys gal_keys;
     keygen_->create_galois_keys(gal_keys);
     return gal_keys;
