+++ pir_bfv/pir_client.h
@@ -44,6 +44,10 @@ public:
     Plaintext replace_element(Plaintext pt, std::vector<std::uint64_t> new_element,
                               std::uint64_t offset);
 
+#ifdef PIR_USE_HARDWARE
+    std::shared_ptr<PoseidonContext> get_context() const { return context_; }
+#endif
+
 private:
     ParametersLiteral enc_params_;
     PirParams pir_params_;
@@ -58,7 +62,9 @@ private:
     std::vector<uint64_t> indices_;  // the indices for retrieval.
     std::vector<uint64_t> inverse_scales_;
 
-    friend class PIRServer;
+#ifdef PIR_USE_HARDWARE
+    friend class PIRServerHardware;
+#endif
 };
 
 }  // namespace pir
